Finding Nash Equilibrium in a Duopoly
Businesses need to change their prices regularly to maximize profits. With data businesses come to know certain empirical relations between their prices and sales. The market is very responsive to the prices set by a certain supplier or seller. The competitors keep a close eye on each other’s prices to be quick in response to the changes in prices. This results in a rather complex interplay of sales and prices of competitors in a market situations.